Fix compatibility with Python 3.10
Python 3.10 requires that PY_SSIZE_T_CLEAN is defined before loading
Python.h, otherwise it will fail at runtime when "s#" is used as a
format string in PyArg_ParseTupleAndKeywords(). Define it and parse
the string length into a Py_ssize_t. It gets assigned to the size_t
explicitly anyway later on.

Closes #1